+++ Engine/Source/Runtime/src/Renderer/Metal/MetalCommandBuffer.mm
@@ -106,12 +106,12 @@ void CommandBuffer::SetViewport(const Size& size) const
     [mHandle->renderCommandEncoder setViewport:viewport];
 }
 
-void CommandBuffer::SetVertexBuffer(const NativeGraphicsHandle buffer, BufferUsage usage, uint32_t size, uint32_t index, uint32_t offset) const
+void CommandBuffer::SetVertexBuffer(const NativeGraphicsHandle buffer, BufferUsage usage, size_t size, uint32_t index, uint32_t offset) const
 {
     [mHandle->renderCommandEncoder setVertexBuffer:buffer offset:offset atIndex:index];
 }
 
-void CommandBuffer::SetFragmentBuffer(const NativeGraphicsHandle buffer, BufferUsage usage, uint32_t size, uint32_t index, uint32_t offset) const
+void CommandBuffer::SetFragmentBuffer(const NativeGraphicsHandle buffer, BufferUsage usage, size_t size, uint32_t index, uint32_t offset) const
 {
     [mHandle->renderCommandEncoder setFragmentBuffer:buffer offset:offset atIndex:index];
 }
